api/requirements.txt
sqlalchemy==1.4.1
pandas==1.2.4
flask==1.1.2
pymysql
pycryptodome
pyjwt
cryptography
